+++ www/go/modules/community/imapauthenticator/language/cs.php
@@ -6,7 +6,7 @@ return array (
   'Enter the domains this imap server should be used to authenticate. Users must login with their e-mail address and if the domain matches this profile it will be used.' => 'Zadejte domény, které má tento IMAP server použít k ověření. Uživatelé se musí přihlásit svou e-mailovou adresou a pokud se doména shoduje s tímto profilem, bude použita.',
   'Validate certificate' => 'Ověření certifikátu',
   'Remove domain from username' => 'Odebrat doménu z uživatelského jména',
-  'Users must login with their full e-mail adress. Enable this option if the IMAP excepts the username without domain.' => 'Uživatelé se musí přihlásit kompletní e-mailovou adresou. Tuto možnost povolte, pokud IMAP nepodporuje uživatelská jména bez domény.',
+  'Users must login with their full e-mail address. Enable this option if the IMAP expects the username without domain.' => 'Uživatelé se musí přihlásit kompletní e-mailovou adresou. Tuto možnost povolte, pokud IMAP nepodporuje uživatelská jména bez domény.',
   'Use authentication' => 'Použít ověřování',
   'Enable this if the SMTP server credentials are identical to the IMAP server.' => 'SMTP přihlašovací údaje jsou stejné jako pro IMAP',
   'User options' => 'Uživatelská nastavení',