+++ CHANGELOG.md
@@ -9,12 +9,18 @@ Master
 - Core: Module dependencies are respected
 - Business: New employee management module
 - History: new module replacing Activity log
+- Comments: Editable date field
 
 end 6.5
 --------
 
 
 
+
+16-11-2020 6.4.192
+- Comments: Fixed bug when adding two comments without opening another item would overwrite the first.
+- Address book: Filter contacts based on a project filter
+- ldapauthenticator: Fixed disabling authentication in LDAP server profile
 
 16-11-2020 6.4.191
 - Email: Fixed problem in Thunderbird e-mails that were cut off. Issue #632.